pyproject.toml
[project]
name = "rsjappmianyang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "ddddocr==1.5.6",
    "pillow>=12.1.1",
    "pyexecjs>=1.5.1",
    "requests>=2.32.5",
]
